Separated review files to folders
